requirements.txt
ccxt
openai
pandas
schedule
python-dotenv
requests
urllib3
flask==3.0.0
flask-cors==4.0.0